LeadDistroProject-Staging_Merge_26_9_2023/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
boto3==1.24.12
botocore==1.27.12
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.6
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2
django-appconf==1.0.5
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-hijack==3.2.1
django-htmx==1.12.0
django-import-export==2.8.0
django-maintenance-mode==0.16.3
django-ses==3.1.0
django-smtp-ssl==1.0
django-tables2==2.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.3
importlib-metadata==1.7.0
IPy==1.1
jmespath==1.0.1
kombu==5.2.4
MarkupPy==1.14
multidict==6.0.2
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-fsutil==0.6.1
pytz==2022.1
PyYAML==6.0
rcssmin==1.1.0
redis==4.3.4
requests==2.27.1
rjsmin==1.2.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.1
zipp==3.8.0
